Cleanup: make x509.LoadCertFromFile public and extract into common
I need to use this routine for the "key ca revoke-device-ca" command.
